requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pymongo[srv]
certifi
